Add stream decoder
