+++ json.go
@@ -264,3 +264,14 @@ func UnmarshalNoEscape(data []byte, v interface{}) error {
 	var dec Decoder
 	return dec.decodeForUnmarshalNoEscape(src, v)
 }
+
+// A Token holds a value of one of these types:
+//
+//	Delim, for the four JSON delimiters [ ] { }
+//	bool, for JSON booleans
+//	float64, for JSON numbers
+//	Number, for JSON numbers
+//	string, for JSON string literals
+//	nil, for JSON null
+//
+type Token interface{}
